# dist/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Sample spec",
        "description": "crud de pruebas"
    },
...
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/obtener-usuarios": {
            "get": {
                "description": "regresa a los usuarios de negocio",
                "operationId": "findById",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "pag",
                    "in": "body",
                    "description": "body",
                    "required": false,
                    "type": "integer"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/getRes"
                        }
                    }
                }
            }
        },
...
        "/user": {
            "get": {
                "description": "regresa al usuario principal",
                "operationId": "encuentra a usuario que inicia session",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "idUsuario",
                    "in": "body",
                    "description": "ID e",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/usuarioAplicacion"

                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "usuarioAplicacion": {
            "type": "object",
            "properties": {
                "nombre": {
                    "type": "string"
                },
                "apellido_paterno": {
                    "type": "string"
                },
                "apellido_materno": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "google": {
                    "type": "string"
                },
                "correo": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "imagen": {

                }
            }
        },
...
        "getRes": {
            "type": "array",
            "items": {
                "type": "object",
                "properties": {
                    "nombre": {
                        "type": "string"
                    },
                    "apellido_paterno": {
                        "type": "string"
                    },
                    "apellido_materno": {
                        "type": "string"
                    },
                    "edad": {
                        "type": "integer"
                    },
                    "correo": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}